+++ spinnaker/provider.go
@@ -1,7 +1,7 @@
 package spinnaker
 
 import (
-	"github.com/hashicorp/terraform/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/spf13/pflag"
 	gate "github.com/spinnaker/spin/cmd/gateclient"
 )
@@ -43,14 +43,16 @@ func Provider() *schema.Provider {
 		},
 		ResourcesMap: map[string]*schema.Resource{
 			"spinnaker_application":       resourceSpinnakerApplication(),
+			"spinnaker_canary_config":     resourceSpinnakerCanaryConfig(),
 			"spinnaker_pipeline":          resourcePipeline(),
 			"spinnaker_pipeline_template": resourcePipelineTemplate(),
 			"spinnaker_project":           resourceSpinnakerProject(),
 		},
 		DataSourcesMap: map[string]*schema.Resource{
-			"spinnaker_application": datasourceApplication(),
-			"spinnaker_pipeline":    datasourcePipeline(),
-			"spinnaker_project":     datasourceProject(),
+			"spinnaker_application":   datasourceApplication(),
+			"spinnaker_canary_config": datasourceCanaryConfig(),
+			"spinnaker_pipeline":      datasourcePipeline(),
+			"spinnaker_project":       datasourceProject(),
 		},
 		ConfigureFunc: providerConfigureFunc,
 	}