+++ spinnaker/resource_application.go
@@ -2,9 +2,10 @@ package spinnaker
 
 import (
 	"fmt"
+	"regexp"
 	"strings"
 
-	"github.com/hashicorp/terraform/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	"github.com/mercari/terraform-provider-spinnaker/spinnaker/api"
 )
 
@@ -21,6 +22,7 @@ func resourceSpinnakerApplication() *schema.Resource {
 				Deprecated:    "use `name` instead",
 				Optional:      true,
 				ConflictsWith: []string{"name"},
+				ValidateFunc:  validateSpinnakerApplicationName,
 			},
 			"name": {
 				Description:  "Name of the Application",
@@ -148,12 +150,12 @@ func resourceSpinnakerApplicationRead(d *schema.ResourceData, meta interface{}) 
 		d.Set("instance_port", v)
 	}
 	if v := app.Attributes.Permissions; v != nil {
-		terraformPermissions, err := buildTerraformPermissions(v)
+		tfPermissions, err := buildTerraformPermissions(v)
 		if err != nil {
 			return err
 		}
 
-		d.Set("permissions", terraformPermissions)
+		d.Set("permissions", tfPermissions)
 	}
 
 	return nil
@@ -277,4 +279,13 @@ func buildTerraformPermissions(permissions *Permissions) (*map[string][]string, 
 	}
 
 	return &users, nil
+}
+
+func validateSpinnakerApplicationName(v interface{}, k string) (ws []string, errors []error) {
+	value := v.(string)
+	if !regexp.MustCompile(`^[a-zA-Z0-9-]+$`).MatchString(value) {
+		errors = append(errors, fmt.Errorf("Only alphanumeric characters or '-' allowed in %q", k))
+	}
+
+	return
 }